Rename position_deal to lifecycle to conform with grid ladders convention
